+++ core/apis_test.go
@@ -27,7 +27,7 @@ import (
 )
 
 func buildTestCoreAPIs(storage core.Storage) *core.APIs {
-	return core.NewCoreAPIs("local", "1.1.1", "build", storage, nil, nil, nil)
+	return core.NewCoreAPIs("local", "1.1.1", "build", "core", storage, nil, nil, nil)
 }
 
 //Services